backend/requirements.txt
fastapi==0.136.0
uvicorn==0.46.0
pydantic>=2.9.4
google-genai==1.73.1
python-dotenv==1.2.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
google-api-core==2.30.3
